prepare image files on change
When the image processing takes more than one second the display
upload timeslot is passed by and for the update the next checkin
slot need to be taken. This delays the image update by one checkin
slot. To avoid that the image files processing will be done by
a polling the image files modification time changes to trigger
on image creation/change.

Signed-off-by: Frank Kunz <mailinglists@kunz-im-inter.net>